+++ src/components/AddressBlock/index.tsx
@@ -24,6 +24,7 @@ import Tooltip from '../Tooltip';
 import TokenNav from '../TokenNav';
 
 import SignVerifyModal from '../../featureModals/SignVerify';
+import AuthModal from '../../featureModals/Auth';
 import { setModalOpen, clearModal } from '../../reduxContent/modal/actions';
 import { changeAccountThunk } from '../../reduxContent/app/thunks';
 import { getSelectedNode } from '../../reduxContent/settings/selectors';
@@ -156,7 +157,8 @@ function AddressBlock(props: Props) {
     const { t } = useTranslation();
     const dispatch = useDispatch();
     const selectedAccountHash = useSelector<RootState, string>(state => state.app.selectedAccountHash);
-    const isSignModalOpen = useSelector<RootState, boolean>(state => state.modal.open);
+    const isModalOpen = useSelector<RootState, boolean>(state => state.modal.open);
+    const activeModal = useSelector<RootState, string>(state => state.modal.activeModal);
     const selectedNode = useSelector(getSelectedNode);
     const tokens = useSelector((state: RootState) => state.wallet.tokens);
 
@@ -194,8 +196,8 @@ function AddressBlock(props: Props) {
         dispatch(changeAccountThunk(addressId, publicKeyHash, index, identityIndex, addressType));
     }
 
-    function setIsSignModalOpen(open) {
-        dispatch(setModalOpen(open));
+    function setIsModalOpen(open, active) {
+        dispatch(setModalOpen(open, active));
         if (!open) {
             dispatch(clearModal());
         }
@@ -255,6 +257,8 @@ function AddressBlock(props: Props) {
         t('components.addressBlock.descriptions.description4')
     ];
     const ready = isReady(status, storeType);
+    const isSignModalOpen = isModalOpen && activeModal === 'sign';
+    const isAuthModalOpen = isModalOpen && activeModal === 'auth';
 
     return (
         <Container>
@@ -307,7 +311,7 @@ function AddressBlock(props: Props) {
             </AddressLabel>
 
             <AddDelegateLabel>
-                <DelegateTitle onClick={() => setIsSignModalOpen(true)}>{t('general.nouns.sign_n_verify')}</DelegateTitle>
+                <DelegateTitle onClick={() => setIsModalOpen(true, 'sign')}>{t('general.nouns.sign_n_verify')}</DelegateTitle>
             </AddDelegateLabel>
 
             <AddDelegateLabel>
@@ -373,7 +377,8 @@ function AddressBlock(props: Props) {
                 );
             })}
 
-            {isSignModalOpen && <SignVerifyModal open={isSignModalOpen} onClose={() => setIsSignModalOpen(false)} />}
+            {isSignModalOpen && <SignVerifyModal open={isSignModalOpen} onClose={() => setIsModalOpen(false, 'sign')} />}
+            {isAuthModalOpen && <AuthModal open={isAuthModalOpen} onClose={() => setIsModalOpen(false, 'auth')} />}
 
             {isInteractModalOpen && (
                 <InteractContractModal open={isInteractModalOpen} onClose={() => setIsInteractModalOpen(false)} addresses={regularAddresses} />
