+++ src/featureModals/Auth/index.tsx
@@ -1,0 +1,3 @@
+import Auth from './Auth';
+
+export default Auth;
