+++ src/reduxContent/modal/types.ts
@@ -1,4 +1,4 @@
 export const SET_MODAL_OPEN = 'SET_MODAL_OPEN';
 export const SET_MODAL_ACTIVE_TAB = 'SET_MODAL_ACTIVE_TAB';
-export const SET_MODAL_TAB = 'SET_MODAL_TAB';
+export const SET_MODAL_VALUE = 'SET_MODAL_VALUE';
 export const CLEAR_MODAL = 'CLEAR_MODAL';
